+++ packages/imperative/src/rest/__tests__/client/__model__/MockHttpRequestResponse.ts
@@ -36,6 +36,10 @@ export class MockHttpRequestResponse extends EventEmitter {
         // doing important stuff with input data here
     }
 
+    public destroy(err?: Error) {
+        if (err) { this.emit("error", err); }
+    }
+
     /**
      * Methods for response
      */
@@ -71,4 +75,9 @@ export class MockHttpRequestResponse extends EventEmitter {
      * @memberof MockHttpRequestResponse
      */
     public reusedSocket: boolean = false;
+
+    /**
+     * Simulate a socket
+     */
+    public socket: any = {};
 }