+++ packages/cli/src/zosfiles/list/ds/DataSet.definition.ts
@@ -52,7 +52,7 @@ export const DataSetDefinition: ICommandDefinition = {
         },
         {
             description: strings.EXAMPLES.EX2,
-            options: `"ibmuser.cntl" -a`
+            options: `"ibmuser.cntl" --attributes`
         },
         {
             description: strings.EXAMPLES.EX3,
@@ -60,7 +60,7 @@ export const DataSetDefinition: ICommandDefinition = {
         },
         {
             description: strings.EXAMPLES.EX4,
-            options: `"ibmuser.*" -a`
+            options: `"ibmuser.*" --attributes`
         },
         {
             description: strings.EXAMPLES.EX5,